requirements.txt
requests
pywebview
pandas
folium
plotly
python-dotenv
shapely
PyQt5
PyQtWebEngine
flask